requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymongo==4.4.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
pydantic==1.10.7
